Working towards GM
NEW:
Finished creating the game logic to represent the "Button" and "Simple wires" modules. These modules now appear (near) complete and include game generation allowing the Arduino to randomly create the scenario for each game, and understands the manual logic to automatically understand the correct and incorrect inputs.
NEW:
Added LCD screen support which now displays information, such as game setup (shows the colours of the simple wires module to aid in preparing the game) and then displays the serial number during the game.
CHANGE:
Changing the timer method which now supports the clock speeding up when a strike is accrued.

BUG:
Button module having trouble with switch state handling. This either sticks in a state where the button is always held down or is always a short press and never a long hold - however this appears due to a memory corruption and a bool "isthisaspress" is either 0 or 1 within check_button() but can return a value in the high 40s or low 50s which disrupts the outcome of the button logic.